merge_release_21_4_ww31_to_master

## Summary

Merge release ww31 to master

## Reviewers

* [ ] @askrebko 
* [ ] @mznamens 

## CI

* [x] [IE-MDK B0 Ubunut18](https://vpux-ci-vpuip.inn.intel.com/job/IE-MDK/job/manual/job/RunTests-KMB/135/)
* [x] [IE-MDK TBH Ubuntu18](https://vpux-ci-vpuip.inn.intel.com/job/IE-MDK/job/manual/job/RunTests-KMB/192/)
* [x] [IE-MDK B0 Windows10](https://vpux-ci-vpuip.inn.intel.com/job/IE-MDK/job/manual/job/RunTests-KMB/190/)
* [x] [Nets-Validation Windows](https://vpux-ci-vpuip.inn.intel.com/job/Nets-Validation/job/manual/job/Windows/6/)
* [x] [Nets-Validation Yocto](https://vpux-ci-vpuip.inn.intel.com/job/Nets-Validation/job/manual/job/Yocto/165/)

See merge request inference-engine/kmb-plugin!3085

+++ src/vpux_compiler/include/vpux/compiler/utils/quantization.hpp
@@ -30,10 +30,10 @@ namespace vpux {
 
 std::tuple<double, int64_t> calcScaleAndZeroPoint(int64_t qMin, int64_t qMax, double rMin, double rMax);
 
-mlir::quant::QuantizedType getQuantizedType(Const::ContentAttr lowConst, Const::ContentAttr highConst, uint32_t levels,
+mlir::quant::QuantizedType getQuantizedType(Const::ContentAttr lowConst, Const::ContentAttr highConst, int64_t levels,
                                             mlir::FloatType realType, mlir::Location loc);
 
-mlir::LogicalResult getFakeQuantParams(mlir::ShapedType qType, uint32_t& levels, mlir::RankedTensorType& attrType,
+mlir::LogicalResult getFakeQuantParams(mlir::ShapedType qType, int64_t& levels, mlir::RankedTensorType& attrType,
                                        mlir::DenseElementsAttr& rMinAttr, mlir::DenseElementsAttr& rMaxAttr,
                                        mlir::Location loc);
